refactor: register 스크린들의 스타일을 별도 파일로 분리
